move some code to other files
