+++ src/scalar.hpp
@@ -22,3 +22,24 @@ torch::Scalar scalar_from_r_ (SEXP x) {
   }
 };
 
+torch::ScalarType scalar_type_from_string(std::string scalar_type) {
+  if (scalar_type == "int32" | scalar_type == "int") {
+    return torch::kInt;
+  } else if (scalar_type == "float64" | scalar_type == "double") {
+    return torch::kDouble;
+  } else if (scalar_type == "float32" | scalar_type == "float") {
+    return torch::kFloat;
+  }
+  Rcpp::stop("scalar not handled");
+}
+
+std::string scalar_type_to_string(torch::ScalarType scalar_type) {
+  if (scalar_type == torch::kInt) {
+    return "int";
+  } else if (scalar_type == torch::kDouble) {
+    return "double";
+  } else if (scalar_type == torch::kFloat) {
+    return "float";
+  }
+  Rcpp::stop("scalar not handled");
+}
